allow to tune the hooks evaluation scope at definition time
While the main goal of this commit is to provide a way to evaluate
blocks in their definition scope (instead of instance_exec'ing them
mandatorily), it is more flexible than that.

The simplest value for scope is a fixed object (e.g. a class)

    define_hook :myhook, scope: MyClass

Or a proc/lambda, which will be passed the callback and the scope
and should return the new scope. This can be used if e.g. you accept
all types of callback and want the procs to be evaluated in their
definition context (as opposed to hooks default behaviour of evaluating
them in the context of the receiver)

Getting the blocks evaluated in their definition scope is achieved
by returning nil for procs:

    define_hook :myhook, scope: lambda { |callback, scope| scope if !callback.proc? }

If you want all the callbacks to be evaluated in the context of a
sandbox:

    define_hook :myhook, scope: lambda { |callback, scope| Sandbox.new(scope) }